Query applicants instead of applications
